Refactor Growth funcs, implement chart tool
Power and Cost growth functions are now set via preset functions, and
are parameterized, with safe defaults. A new tab has been added to the
Tool Editor, with a chart plotting the growth functions. New fields have
been added to set the growth function type, and the coefficients of the
function. Some fields had to be shuffled around in order for everything
to fit, and some new input components were added to create a more
unified look-and-feel.